code reorganization, add docs for scout
