+++ docs/source/models.rst
@@ -1,0 +1,129 @@
+Models
+======
+
+Base Models and Mixins
+----------------------
+
+Model
+"""""
+
+.. autoclass:: purchasing.database.Model
+    :members:
+
+RefreshSearchViewMixin
+""""""""""""""""""""""
+
+.. autoclass:: purchasing.database.RefreshSearchViewMixin
+    :members:
+
+
+Scout and Conductor
+-------------------
+
+Contracts
+"""""""""
+
+.. autoclass:: purchasing.data.contracts.ContractBase
+    :members:
+
+.. autoclass:: purchasing.data.contracts.ContractType
+    :members:
+
+.. autoclass:: purchasing.data.contracts.ContractProperty
+    :members:
+
+.. autoclass:: purchasing.data.contracts.LineItem
+    :members:
+
+Companies
+"""""""""
+
+.. autoclass:: purchasing.data.companies.Company
+    :members:
+
+.. autoclass:: purchasing.data.companies.CompanyContact
+    :members:
+
+Searches
+""""""""
+
+.. automodule:: purchasing.data.searches
+    :members:
+
+Stages
+""""""
+
+.. autoclass:: purchasing.data.stages.Stage
+    :members:
+
+.. autoclass:: purchasing.data.stages.StageProperty
+    :members:
+
+Flows
+"""""
+
+.. autoclass:: purchasing.data.flows.Flow
+    :members:
+
+Contract Stages
+"""""""""""""""
+
+.. autoclass:: purchasing.data.contract_stages.ContractStage
+    :members:
+
+.. autoclass:: purchasing.data.contract_stages.ContractStageActionItem
+    :members:
+
+Beacon
+------
+
+Opportunities
+"""""""""""""
+
+.. autoclass:: purchasing.opportunities.models.Opportunity
+    :members:
+
+.. autoclass:: purchasing.opportunities.models.OpportunityDocument
+    :members:
+
+.. autoclass:: purchasing.opportunities.models.RequiredBidDocument
+    :members:
+
+Vendors
+"""""""
+
+.. autoclass:: purchasing.opportunities.models.Vendor
+    :members:
+
+Categories
+""""""""""
+
+.. autoclass:: purchasing.opportunities.models.Category
+    :members:
+
+Public and User
+---------------
+
+Users
+"""""
+
+.. autoclass:: purchasing.users.models.User
+    :members:
+
+.. autoclass:: purchasing.users.models.Role
+    :members:
+
+.. autoclass:: purchasing.users.models.Department
+    :members:
+
+App Status
+""""""""""
+
+.. autoclass:: purchasing.public.models.AppStatus
+    :members:
+
+Job Status
+""""""""""
+
+.. autoclass:: purchasing.jobs.job_base.JobStatus
+    :members:
