--- docs/source/technical.rst
@@ -1,14 +1,0 @@
-Technical Docs
-==============
-
-:doc:`/beacon`
---------------
-
-:doc:`/scout`
---------------
-
-:doc:`/conductor`
------------------
-
-:doc:`/public`
---------------
