+++ purchasing/scout/views.py
@@ -11,13 +11,16 @@ from flask_login import current_user
 from purchasing.database import db
 from purchasing.utils import SimplePagination
 from purchasing.decorators import wrap_form, requires_roles
-from purchasing.notifications import Notification
 
-from purchasing.scout.forms import SearchForm, FeedbackForm, NoteForm
-from purchasing.users.models import Department, User, Role
-from purchasing.data.searches import SearchView, find_contract_metadata, return_all_contracts
+from purchasing.scout.forms import SearchForm, NoteForm
+from purchasing.users.models import Department
 from purchasing.data.companies import Company
 from purchasing.data.contracts import ContractBase, ContractNote, ContractType
+
+from purchasing.scout.util import (
+    build_filter, build_cases, feedback_handler,
+    find_contract_metadata, return_all_contracts, FILTER_FIELDS
+)
 
 from purchasing.scout import blueprint
 
@@ -26,15 +29,17 @@ CRAZY_CHARS = re.compile('[^A-Za-z0-9 ]')
 @blueprint.route('/', methods=['GET'])
 @wrap_form(SearchForm, 'search_form', 'scout/explore.html')
 def explore():
-    '''
-    The landing page for scout. Renders the "big search"
-    template.
+    '''The landing page for scout. Renders the "big search" template.
+
+    :status 200: Renders the appropriate landing page.
     '''
     return dict(current_user=current_user, choices=Department.choices())
 
 @blueprint.route('/filter', methods=['GET'])
 def filter_no_department():
     '''The landing page for filtering by departments
+
+    :status 200: Renders the appropriate landing page.
     '''
     return render_template(
         'scout/filter.html',
@@ -48,89 +53,76 @@ def filter_no_department():
 
 @blueprint.route('/filter/<int:department_id>', methods=['GET'])
 def filter(department_id):
+    '''Filter contracts by which ones have departmental subscribers
+
+    :param department_id: Department's unique ID
+    :status 200: Renders template with all contracts followed by that department
+    :status 404: When department is not found with the specified ID
     '''
-    Filter contracts by which ones have departmental subscribers
-    '''
-    pagination_per_page = current_app.config.get('PER_PAGE', 50)
-    page = int(request.args.get('page', 1))
-    lower_bound_result = (page - 1) * pagination_per_page
-    upper_bound_result = lower_bound_result + pagination_per_page
     department = Department.query.get(int(department_id))
 
-    contracts = db.session.execute(
-        '''
-        SELECT
-            contract.id, description,
-            count(contract_user_association.user_id) AS follows
-        FROM contract
-        LEFT OUTER JOIN contract_user_association
-            ON contract.id = contract_user_association.contract_id
-        FULL OUTER JOIN users
-            ON contract_user_association.user_id = users.id
-        JOIN department
-            ON users.department_id = department.id
-        WHERE department.id = :department
-        GROUP BY 1,2
-        HAVING count(contract_user_association.user_id) > 0
-        ORDER BY 3 DESC, 1 ASC
-        ''', {'department': int(department_id)}
-    ).fetchall()
+    if department:
+        pagination_per_page = current_app.config.get('PER_PAGE', 50)
+        page = int(request.args.get('page', 1))
+        lower_bound_result = (page - 1) * pagination_per_page
+        upper_bound_result = lower_bound_result + pagination_per_page
 
-    if len(contracts) > 0:
-        pagination = SimplePagination(page, pagination_per_page, len(contracts))
-        results = contracts[lower_bound_result:upper_bound_result]
-    else:
-        pagination = None
-        results = []
+        contracts = db.session.execute(
+            '''
+            SELECT
+                contract.id, description,
+                count(contract_user_association.user_id) AS follows
+            FROM contract
+            LEFT OUTER JOIN contract_user_association
+                ON contract.id = contract_user_association.contract_id
+            FULL OUTER JOIN users
+                ON contract_user_association.user_id = users.id
+            JOIN department
+                ON users.department_id = department.id
+            WHERE department.id = :department
+            GROUP BY 1,2
+            HAVING count(contract_user_association.user_id) > 0
+            ORDER BY 3 DESC, 1 ASC
+            ''', {'department': int(department_id)}
+        ).fetchall()
 
-    current_app.logger.info('WEXFILTER - {department}: Filter by {department}'.format(
-        department=department.name
-    ))
+        if len(contracts) > 0:
+            pagination = SimplePagination(page, pagination_per_page, len(contracts))
+            results = contracts[lower_bound_result:upper_bound_result]
+        else:
+            pagination = None
+            results = []
 
-    return render_template(
-        'scout/filter.html',
-        search_form=SearchForm(),
-        results=results,
-        pagination=pagination,
-        department=department,
-        choices=Department.choices(),
-        path='{path}?{query}'.format(
-            path=request.path, query=request.query_string
+        current_app.logger.info('WEXFILTER - {department}: Filter by {department}'.format(
+            department=department.name
+        ))
+
+        return render_template(
+            'scout/filter.html',
+            search_form=SearchForm(),
+            results=results,
+            pagination=pagination,
+            department=department,
+            choices=Department.choices(),
+            path='{path}?{query}'.format(
+                path=request.path, query=request.query_string
+            )
         )
-    )
-
-def build_filter(req_args, fields, search_for, filter_form, _all):
-    '''Build the where clause filter
-    '''
-    clauses = []
-    for arg_name, _, filter_column in fields:
-        if _all or req_args.get(arg_name) == 'y':
-            if not _all:
-                filter_form[arg_name].checked = True
-            clauses.append(filter_column.match(
-                search_for,
-                postgresql_regconfig='english')
-            )
-    return clauses
-
-def build_cases(req_args, fields, search_for, _all):
-    '''Build the case when statements
-    '''
-    clauses = []
-    for arg_name, arg_description, filter_column in fields:
-        if _all or req_args.get(arg_name) == 'y':
-            clauses.append(
-                (filter_column.match(
-                    search_for,
-                    postgresql_regconfig='english'
-                ) == True, arg_description)
-            )
-    return clauses
+    abort(404)
 
 @blueprint.route('/search', methods=['GET', 'POST'])
 def search():
-    '''
-    The search results page for scout
+    '''The search results page for scout
+
+    In order to create permalinks to each search/filter result combination,
+    POST methods have their form arguments popped off and then are immediately
+    redirected to GET methods.
+
+    .. seealso ::
+        * :py:mod:`purchasing.data.searches` for more on the search query
+        * :py:class:`purchasing.scout.forms.SearchForm` for the search form construction
+        * :py:func:`purchasing.scout.util.build_filter` for how filters are built
+        * :py:func:`purchasing.scout.util.build_cases` for how case statements are built
     '''
     if request.method == 'POST':
         args = request.form.to_dict()
@@ -156,18 +148,9 @@ def search():
     lower_bound_result = (page - 1) * pagination_per_page
     upper_bound_result = lower_bound_result + pagination_per_page
 
-    # build filter and filter form
-    fields = [
-        ('company_name', 'Company Name', SearchView.tsv_company_name),
-        ('line_item', 'Line Item', SearchView.tsv_line_item_description),
-        ('contract_description', 'Contract Description', SearchView.tsv_contract_description),
-        ('contract_detail', 'Contract Detail', SearchView.tsv_detail_value),
-        ('financial_id', 'Controller Number', SearchView.financial_id),
-    ]
-
     filter_or = build_filter(
-        request.args, fields, search_for, search_form,
-        not any([request.args.get(name) for name, _, _ in fields])
+        request.args, FILTER_FIELDS, search_for, search_form,
+        not any([request.args.get(name) for name, _, _ in FILTER_FIELDS])
     )
 
     filter_and = []
@@ -178,8 +161,8 @@ def search():
         search_form.contract_type.data = ContractType.query.get(int(request.args.get('contract_type')))
 
     found_in_case = build_cases(
-        request.args, fields, search_for,
-        not any([request.args.get(name) for name, _, _ in fields])
+        request.args, FILTER_FIELDS, search_for,
+        not any([request.args.get(name) for name, _, _ in FILTER_FIELDS])
     )
 
     # determine if we are getting archived contracts
@@ -225,6 +208,12 @@ def search():
 @blueprint.route('/companies/<int:company_id>')
 @wrap_form(SearchForm, 'search_form', 'scout/company.html')
 def company(company_id):
+    '''Company profile page
+
+    :param contract_id: Unique ID for a :py:class:`purchasing.data.company.Company` object
+    :status 200: Renders the company profile template
+    :status 404: Unique company ID not found
+    '''
     company = Company.query.get(company_id)
 
     if company:
@@ -242,6 +231,14 @@ def company(company_id):
 @wrap_form(SearchForm, 'search_form', 'scout/contract.html')
 def contract(contract_id):
     '''Contract profile page
+
+    For GET requests, render the profile page. For POSTs,
+    try to submit a new note.
+
+    :param contract_id: Unique ID for a :py:class:`purchasing.data.contracts.ContractBase` object
+    :status 200: Renders the contract profile template
+    :status 302: Try to post note, then redirect to the same page's contract view
+    :status 404: Unique contract ID not found
     '''
 
     contract = ContractBase.query.get(contract_id)
@@ -276,74 +273,42 @@ def contract(contract_id):
         )
     abort(404)
 
-def feedback_handler(contract_id=None, search_for=None):
-    '''
-    Allow user to send feedback on the data present in a specific contract
-    '''
-    if contract_id:
-        contract = ContractBase.query.get(contract_id)
-    else:
-        contract = ContractBase(
-            description='Search term: ' + search_for
-        )
-
-    search_form = SearchForm()
-    if contract:
-        form = FeedbackForm()
-
-        if not current_user.is_anonymous():
-            form.sender.data = current_user.email
-
-        if form.validate_on_submit():
-
-            current_app.logger.info('WEXFEEDBACK - Feedback from {email} about {contract}'.format(
-                email=form.sender.data,
-                contract=contract.description
-            ))
-
-            feedback_sent = Notification(
-                to_email=db.session.query(User.email).join(Role, User.role_id == Role.id).filter(
-                    Role.name.in_(['admin', 'superadmin'])
-                ).all(),
-                subject='Scout contract feedback - ID: {id}, Description: {description}'.format(
-                    id=contract.id if contract.id else 'N/A',
-                    description=contract.description
-                ), html_template='scout/feedback_email.html',
-                contract=contract, sender=form.data.get('sender'),
-                body=form.data.get('body')
-            ).send()
-
-            if feedback_sent:
-                flash('Thank you for your feedback!', 'alert-success')
-            else:
-                flash('Oh no! Something went wrong. We are looking into it.', 'alert-danger')
-
-            if contract.id:
-                return redirect(url_for('scout.contract', contract_id=contract.id))
-            return redirect(url_for('scout.explore'))
-
-        return render_template(
-            'scout/feedback.html',
-            search_form=search_form,
-            contract=contract,
-            choices=Department.choices(),
-            feedback_form=form,
-            search_for=search_for
-        )
-    abort(404)
-
 @blueprint.route('/contracts/feedback/<search_for>', methods=['GET', 'POST'])
 def search_feedback(search_for):
-    return feedback_handler(search_for=search_for)
+    '''Provide feedback about an empty search
+
+    This page is only viewable in the event of a search that returns 0 results.
+
+    :param search_for: Search term.
+
+    .. seealso ::
+        :py:mod:`purchasing.scout.util.feedback_handler` for information on how
+        the feedback is processed and handled
+    '''
+    contract = ContractBase(description='Search term: ' + search_for)
+    return feedback_handler(contract, search_for=search_for)
 
 @blueprint.route('/contracts/<int:contract_id>/feedback', methods=['GET', 'POST'])
 def feedback(contract_id):
-    return feedback_handler(contract_id=contract_id)
+    '''Provide feedback about a contract
+
+    :param contract_id: Unique ID for a :py:class:`purchasing.data.contracts.ContractBase` object
+
+    .. seealso ::
+        :py:mod:`purchasing.scout.util.feedback_handler` for information on how
+        the feedback is processed and handled
+    '''
+    contract = ContractBase.query.get(contract_id)
+    if contract:
+        return feedback_handler(contract=contract)
+    abort(404)
 
 @blueprint.route('/contracts/<int:contract_id>/subscribe')
 @requires_roles('staff', 'admin', 'superadmin', 'conductor')
 def subscribe(contract_id):
     '''Subscribes a user to receive updates about a particular contract
+
+    :param contract_id: Unique ID for a :py:class:`purchasing.data.contracts.ContractBase` object
     '''
     contract = ContractBase.query.get(contract_id)
     next_url = request.args.get('next', '/scout')
@@ -371,6 +336,8 @@ def subscribe(contract_id):
 @requires_roles('staff', 'admin', 'superadmin', 'conductor')
 def unsubscribe(contract_id):
     '''Unsubscribes a user from receiving updates about a particular contract
+
+    :param contract_id: Unique ID for a :py:class:`purchasing.data.contracts.ContractBase` object
     '''
     contract = ContractBase.query.get(contract_id)
     next_url = request.args.get('next', '/scout')
